fix(#5): Skip multi schema tests for oracle DB
Oracle DB requires user in query to drop a particular schema.
currently SQLA seems to generate a query
	`DROP schema schema_name`
but Oracle requires
	`DROP user schema_name`

Tried sqla.schema.DropSchema also still it is not working so skipping
as teardown is not working and remaining tests start failing.

For now, we won't support multi schema testing with oracle (also not
sure if the package works in that scenario)