fix(#77): Stop assuming ordered txn-id in versions
The versions attribute was trying to find all version records for
a single model. But it assumed that the transaction_ids will be
ordered.

Change the `order_by` clause to a subquery.
NOTE: This is not very performant. But the simplest way to do this.
There is an alternate approach documented using a non-primary mapper
https://groups.google.com/g/sqlalchemy/c/cQ_Y2gJWj28
But non-primary mappers are discontinued in favor of aliased().
and when I tried it with aliased() I was getting errors.