fix(#77): Set txn-id in sqlalchemy to be incremental
In sqlalchemy, ensure that the txn-id will always be incrementing.
This is an assumption that is baked into the library and so we should
ensure this assumption never breaks for txn-id atleast.

This can lead to a small chance of a race condition because we are not
locking the DB while doing this ... Which we need to assess and see it's
impact.