fix(39): set end_transaction_id for association columns as well
since we are in direction to support versioning control on tables
as well as ORM models, extended support for filling end transaction
id column in association table as well when manager has validity
strategy enabled.
this for now is a kind of hacky as we don't support specifying
configurations on table level so we refer configurations mentioned
in manager and apply it globally, alternatively we can give an
option to disable this behaviour even though validity strategy is
enabled but didn't encorporated it because it seemed bit of an
over-configgyy solution for now want to let the value be set.